Fixed issue with Flood losses not populating. Updated EQ deterministic Downloads Category to be Results.
Turned returnPeriod into downloadCategory variable to have more control on its value for different perils. Added upper bounds to losses to resolve issue with units locking at thousands. Adjusted losses for USD instead of thousands as the getTotalEconomicLoss provides in USD and not thousands.  Had to move the losses to inside the downloads section since Flood lossess query relies on the returnperiod value as a filter.